requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cramjam==2.10.0
discord.py==2.5.2
exllamav2==0.3.1
fastparquet==2024.11.0
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
Flask==2.2.5
